Removing the Kmeans libraries from this folder. Moved them to utils. Additionally these files were preped for the removal of the UAS compile. I cleaned up this code to be slightly more efficient. Combines CSpaceClearance and Clearance features in the same file. Cleaned up XML input items, classes, unused variables, eliminated warnings etc.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@2839 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Partitioning/Partition.cpp
@@ -21,7 +21,7 @@ Partition::~Partition(){
 BoundingBox Partition::GetBoundingBox(){
   BoundingBox bb(3,3);
   double minX=1e6, maxX=-1e6, minY=1e6, maxY=-1e6, minZ=1e6, maxZ=-1e6;
-  for(int i=0; i<m_vvid.size() ; i++){
+  for(size_t i=0; i<m_vvid.size() ; i++){
      CfgType cfg = m_rdmp->m_pRoadmap->find_vertex(m_vvid[i])->property();
     vector<double> position = cfg.GetPosition();
     if(position[0]<minX)minX=position[0];
